Implement DH_KABOXER_BUILD_STRATEGY
The package build daemons don't have a running docker service so we
can't really call "docker build" there. So tweak dh_kaboxer to not
call "kaboxer build" by default (assuming that the image build happens
elsewhere and that we just register the image location in kaboxer.yaml)
except if the build strategy is to build a tarball to embed in the
package.

This obsoletes DH_KABOXER_BUILD_TARBALL=1 which is replaced by
DH_KABOXER_BUILD_STRATEGY=tarball.