Some more fixes in README
In existing applications, `<app>.kaboxer.yaml` is used, rather than
`debian/<package>.kaboxer.yaml`. Note that the name itself does not
matter that much, as kaboxer look for files named as such:

    'kaboxer.yaml', '*.kaboxer.yaml'

and then rely on the id that is found in the file.

What matters however is that the file should be in the root directory,
and not in the debian/ directory.