Add test to ensure 'kaboxer purge non-existing-app' succeeds
This test is supposed to catch the issue mentioned at #55, but it does
not. The issue #55 is somewhat more complicated, it is due to the fact
that the image 'kaboxer/joplin' does not exist, however the image
'registry.gitlab.com/kalilinux/packages/joplin-kbx/kaboxer/joplin'
exists.

It is not worth introducing a test for that particular case though,
since the idea is to get rid of the name that includes the registry,
as proposed in #68.

However it does not hurt to add a test to ensure that running the purge
command on a non-existing image succeeds (ie. returns 0).

(Not saying this is the right behavior, but it's how it behaves at the
moment, so let's add a test to make sure this behavior does not change
unexpectedly.)