run: Simplify how detach opt is handled
Looking carefully at the code, there is no need to set opts['detach'] so
early. There are only two places where it's used:
- as part of opts2 (copy from opts), but then it's set to False
  unconditionnaly.
- then at the very end of the code

So let's move it at the bottom of the function for clarity.

It also matters for the next commit to come, where we will have to do
this test anyway.