dh-kaboxer: Be more verbose in the postinst script
At the moment "kaboxer prepare" is very quiet, there is no output at
all. However it might take minutes to complete, in case it fetches a
docker image from a remote repository.  Not great for user, who have
no clue that there's an operation in progress,  and might think that
kaboxer hangs. Even worse, when users are installing a -kbx package,
then might not even know what kaboxer is,  and think that apt is the
one which hangs.

This commit improves the situation a tiny bit. By adding the flag -v,
apt/kaboxer output now looks like that:

  $ sudo dpkg -i joplin-kbx_1.4.19_all.deb
  (Reading database ... 238077 files and directories currently installed.)
  Preparing to unpack joplin-kbx_1.4.19_all.deb ...
  Unpacking joplin-kbx (1.4.19) over (1.4.19) ...
  Setting up joplin-kbx (1.4.19) ...
  Preparing joplin
  Pulling registry.gitlab.com/kalilinux/packages/joplin-kbx/kaboxer/joplin:latest image from registry
  Processing triggers for gnome-menus (3.36.0-1) ...
  Processing triggers for desktop-file-utils (0.26-1) ...
  Processing triggers for mailcap (3.68) ...