Clarification regarding Docker Registry HTTP API v2
The "alternate api" was in fact the most standard API we can get, since
it was standardized by the Open Containers Initiative (OCI). It's just
unfortunate that it's not usable without authentication.

A non-exhaustive list of registries that implements the OCI Distribution
Spec (aka. the Docker Registry HTTP API v2) can be found at:
<https://conformance.opencontainers.org/distribution-spec/>