Rework, assume that the registry url is... the registry url
It might sound weird, but the fact is that before this commit, we
assumed that the registry url was the registry url + an optional path,
and that this path was a part of the image name. More specifically, for
GitLab, we assumed that this optional path was in fact the project path
(aka namespace/project, for example 'kalilinux/packages/covenant-kbx').

That's a lot of (undocumented) assumptions.

So here we go back to something more simple: the registry url, which
comes from the kaboxer.yaml, is in fact the registry url. No more no
less.

Building upon that, we can see that the implementation becomes somewhat
simpler, except when it comes to talk to the GitLab API. This API
requires that we know the "project path", which is a GitLab concept. We
can guess it from the image name, thanks to the fact that GitLab as some
well-defined naming conventions for container images [1]:

  > Images follow this naming convention:
  > <registry URL>/<namespace>/<project>/<image>

Note that:
- the "project path" is in fact <namespace>/<project>
- the <image> part is optional (this is not very clear in the doc)

The downside is that we need to make successive requests against GitLab
until we find the project path. In this (naive) implementation, we start
by assuming that the project path is the image name that is provided in
the kaboxer.yaml file. If it does not work, we remove the last element
and try again, until we find the project path.

It works, but it's sub-optimal, and not very elegant.

This can be mitigated by agreeing on some conventions. We can agree that
we always have an <image> part for our container name (that's in line
with what's written in the GitLab doc, after all), and that this image
name does not contain any slash (I don't see any reason why we'd want to
have slashes in the image name anyway). Then in kaboxer, we can optimize
for this situation, by trying at first to remove the last part of the
image name, and make an API request against that.

If the person writting the kaboxer.yaml follows this convention, kaboxer
will get it right at the first try. Otherwise, it will take some
iterations.

[1] https://docs.gitlab.com/ce/user/packages/container_registry/#image-naming-convention