Optimize listing image tags for GitLab registry
This is based on the assumption that the most likely image names, in the
context of "Docker images for Kaboxer apps hosted on a GitLab registry",
will be <project-path>/<image>, <image> being the application name.

If we have different apps under a project, no problem, that translates
to different image names. For example, the hello example is made of 4
apps, so that's 4 container images:

    kalilinux/tools/kaboxer/hello-{allinone,cli,gui,server}

If we follow this simple convention, ie. that the image name is equal to
the app name, then the "try and guess" code that lists the tags in a
GitLab registry will get it right at the first try.

If one does not follow this convention, either by putting a slash in the
image name (eg. 'hello/my/app'), or by not giving any image name at all
(GitLab allows an image to exist under the name <project-path> without
any additional component), no worries, things still work, but the code
won't find the project path at the first try.