load_config: Drop app=None, make app a positional argument
A value of None for app in this method does not make much sense. If you
look at the code, this method is about picking a yaml config file that
is named either 'app.kaboxer.yaml' either 'kaboxer.yaml', and for which
the value of 'application:id' is equal to app.

If we allow app to be None here, it means that we look for a file named
'kaboxer.yaml' and for which 'application:id' is not defined. Is there
really a use-case for that?

Additionally, as of now this code path is not triggered, ie.
load_config() is never called with app=None, as far as I can see.

So let's just drop this possibility.