cmd_install: Create a separate install_app() method to install each app
Now cmd_install() only runs the loop on available app configs, and for
each config, the app is installed with install_app(). Hence we can
reduce the indent level and make everything a bit more readable.

This is similar to what's done already with the cmd_build() and
build_image() methods.