cmd_build: Move save/push a level up, in cmd_build() rather than build_image()
These two actions (save/push) are post-build steps, so it feels right to
keep it outside of the build_image() method.

By digging these calls up, it also becomes easier to spot a problem in
the current design: in cmd_build(), the method 'self.push()' is called
from within a loop on parsed_configs. However, looking at the actual
implementation of the push() method: this method also loops on
parsed_configs.

This will be fixed in next commits.