cmd_push: Run the loop, and pass parsed_config to push() method
Similar to previous commit. This allows to have the loop in the
top-level method (cmd_push()), and then doing the actual job in the
push() method, hence removing one indent level.

Note that push() was used also from cmd_build(), in a sloppy way. Now
that we have a parsed_config argument to the push() method, we
automatically get rid of the redundant loop that was in effect when
calling push() from build_image().