gitlab-ci: Update the default remote image name
This is to reflect changes made a few weeks ago, where we started
assuming that the image name would be equal to the kaboxer app id.
This assumption makes listing images on a remote GitLab registry a
bit faster.

See 3d57cb7a0dc27c007cc8fc8b92cdf16a4934fedb for some explanations
related to this change.

Note that the right thing to do here would be to call 'kaboxer push'
instead of doing the push manually. We'll get there eventually...