Add a tox job to GitLab CI
Thid job runs tox, and it excludes the unit tests. This is because
unit-tests are already run by the 'build' job (ie. dpkg-buildpackage).
We don't want to run it again here.

However, what we want is to run the code QA tools (flake8 and black),
that are NOT run by dpkg-buildpackage.