Make logger a global object
Having logger belonging to the Kaboxer object makes things difficult.

1) Can't use it outside of the Kaboxer object, so every object must
   define (and configure) their own logger. This leads to
   ContainerRegistry object having its own logger for example. However,
   configuring the logger is left to Kaboxer.setup_logging(), because
   it needs to be done after parsing arguments, and because I didn't
   want to copy/paste this boilerplate in ContainerRegistry.

2) Unit tests are also inmpacted by that, because they also need to
   define the logger object for each object that is tested. This is
   due to the fact that the logger is not created in the __init__
   method of Kaboxer, but instead by the setup_logging() method.
   This leads to awkward things, for example this, in unit_tests.py:

     class TestKaboxerFindConfigsInDir(unittest.TestCase):
         def setUp(self):
             self.obj = Kaboxer()
             import logging
             self.obj.logger = logging.Logger('kaboxer')

There are surely many ways to deal with that and improve the situation.
To me, the most simple is just to take the logger out of any object, and
make it accessible to the whole file, using something akin to what's
recommended by the Python documentation [1]:

    logging.getLogger(__name__)

Execpt that right now, since everything is in the __init__.py file, I'd
rather use 'kaboxer' instead of '__name__'.

This change makes everything simpler, more concise, and we can get rid
of the few awkward hacks added in previous commits:

- 57688969a9a9904a81741d027ba4716fbc4c9240
- ed961d9b8d39dd74e6f5902c6cb5e7b48dc72db4

[1]: https://docs.python.org/3/library/logging.html