Flake8: Ignore A002 argument 'input' is shadowing a python builtin
In this case, the methods are little more than wrappers on top of
subprocess.run(), and our 'input' argument is passed as is to
subprocess.run() 'input' argument. It seems to me that it's the best
name we can give to this arg. Not really our fault if subprocess.run()
has an argument named 'input'.