Drop component from helper filenames if there's only one component
Since cli helpers are user-facing (they are meant to be used in the
terminal by users), let's try to polish it a bit.

If there's only one component in an app (and it could be that it's the
majority of kaboxer apps), then let's drop the component from the helper
filename. In such case the name of the component doesn't matter.

So if we take the example of zenmap, and assuming you're an user wanting
to start it from the terminal.

Before this change, you'd type:

    $ zenmap-default-kbx

Now, it is simply:

    $ zenmap-kbx

This change adds a bit of complexity in the code, but I believe that the
improvment for the user experience is worth it.

Note that we could also do that for desktop files, however desktop files
are not really user-facing (nobody goes and looks for files in
/usr/share/applications), they are instead files loaded by the desktop.
So the name for this files is not so important, let's keep it simple.