Add aliases: start=run, ls=list
To run a kaboxer app in the foreground, `kaboxer run` seems natural. It
is reminiscent of `docker run` or `flatpak run`. Nothing wrong with it.

However kaboxer also allows to run apps in the background (for example,
servers), that will need to be stopped later on with `kaboxer stop`. For
this kind of applications, the pair of commands `kaboxer start | stop`
seems more natural than `kaboxer run | stop`. It is reminiscent of
`systemctl start | stop`.

So this commit adds an alias `start` to the `run` command.

It's worth noting that the distinction between `run` and `start` does
exist and is meaningful for other tools (like docker or lxc). So I hope
we won't regret this choice down the road, and realize that we also need
to distinguish between those two actions. But I don't think so.  Kaboxer
doesn't aim to give users full control of containers, it aims to allow
user to run their applications.

The alias list=ls does not need any explanation I guess.

Fixes: #52