create_xauth(): Rewrite without pipes
The code is quite straightforward to read: create_xauth() is always run
before running a container. create_xauth() always create the file
~/.docker.xauth. And yet, if this file was not present when we started,
the creation of the container fails with:

    docker.errors.APIError: 400 Client Error: Bad Request
    ("invalid mount config for type "bind": bind source path does not exist: /home/kali/.docker.xauth")

So it seems to me that the issue comes from how the two processes are
piped together, and it seems that the final lines g.close() and
f.close() are not enough to ensure that the processes are terminated and
that all the data has been written. Hence when docker tries to run, the
file does not exist yet.

To fix it, it could be as simple as adding a f.wait() and g.wait().
However I'm not so familiar with subprocess and piping.

It also seems even better to NOT pipe anything, in this case we can just
run the two processes one after another. It's more readable to me, and
solves any sync issue.

Fixes: #93