tests: run_command captures output by default
It turns out that the vast majority of the calls to run_command()
either:
- had no output already (ie. silent commands)
- explicitly required to be silent with 'ignore_output=True'

So making run_command() silent by default makes more sense. And since we
make it silent by capturing the output, it also means that from now on,
the command output is available for the caller, if needed.