Rework docker connection setup
This commit doesn't do much, it mostly moves the docker_is_working()
function within the docker_setup() function. It doesn't make much sense
to have both separate, as we don't really handle the case where yes the
connection is setup, but no it's not working.

Additionally this distinction disappeared with docker-py 4.3, as from
now on docker.from_env() fails if we don't have the permissions to talk
to the docker daemon (was not the case before).

By merging these two functions together, the logic becomes simpler, and
by handling any exception that might be raised from docker_setup(), we
naturally fix #101.

Fixes: #101