Drop --reuse-container, use information from kaboxer.yaml instead
This is how it should have been done from the beginning. The setting
'reuse-container' is part of the YAML file, and shouldn't be needed on
the command-line.

For example, for kaboxed Firefox, we don't want to create a new Firefox
instance every time we open a new web page. Instead we want to reuse the
existing Firefox container instance, and open a new tab. That's why we
set "reuse_container" in the YAML file. And from this point, we don't
want user to ask themselves "Should I run with --reuse-container or
not?". Nope, the developer who packaged Firefox for Kaboxer already
answered this question. The decision is not in the hands of the user.

With all of that being said, so what does this commit do?

* 'kaboxer run' now queries the YAML file to know whether it should
  reuse an existing container.
* the generated desktop files and cli helpers don't set the argument
  '--reuse-container' on the command line anymore, which allows for a
  nice simplification in the code.

Additionally, and despite everything said above, we keep a
'--reuse-container' flag on the command-line, HOWEVER:

* this is for development/debug ease of use, therefore it is a hidden
  setting that does not appear in the usage message.
* it has to be invoked as '--reuse-container' or '--no-reuse-container'.
* obviously it takes precedence over the setting in the YAML file.

Fixes: #89