tests: Set iptables legacy only for GitLab CI
We "guess" that we're on GitLab CI simply based on the state of docker.
On GitLab CI it fails to start because of iptables.

If we're not in GitLab CI, then there's no need to set iptables to
legacy.

This is a hopefully a temporary hack.

This is a follow-up of b21d014b73ff20eac87287a7d63fdd3c1a923226