Mount icons in usr/local/ rather than usr
The directory /usr/share/icons/ in the container might contain essential
files for the applications (eg. custom icons that are needed for the
UI), so we don't want to replace it with /usr/share/icons from the host.

Therefore we prefer to use /usr/local/share/icons/. Of course it's
possible that an application ship files in usr/local/share/icons, but
in this case it can be changed at build time, in the Dockerfile, by
setting the prefix to /usr rather than /usr/local. It seems like a
reasonable limitation.