Fix support for current,latest versions
Reported at https://bugs.kali.org/view.php?id=8160. It broke when
python3-packaging 22.0-2 landed in Kali rolling.

This commit is just a workaround, it just makes sure that the current
and latest versions are handled as they were before, and it seems that
it has never been handled correctly.