Fix test suite
In latest versions, Git became picky with file ownership, and refuses to
work when the ownership is weird (like, the repo doesn't belong to the
user, which is the case here).

Fix that, and move the git init code in d/t/run-tests, first because
it's easier to read and write it there, but also because I have the
impression that tests (might) run in parallel, and having the first
running test setup the git repo seems a bit racy to me.